light.inc.frag: Use worldView parameter instead of worldEye
The phongFunction used in `phongDataDefined.frag` uses the
worldPosition parameter. This parameter is then used in
adsModel. However, it turns out that eyePosition is used instead of
worldView. However, the result is correct because adsModel expects the
eyePosition.

This changes ensures that the worldView parameter is used in
phongFunction and adsModel. This way, it helps avoid any confusion and
this is also how these different functions are used in the QT3D source
code.