qgs3dmapscene: Respect 3d rotation when 2D and 3D navigation are sync
Once sync is activated between the 2D view and a 3D view, zooming in
or out from the 2D view resets the rotation in the 3D view. This is
because the `setViewFromTop` method is used to set the view from the
2D extent. This function resets the heading and pitch angles to zero.

This issue is fixed by using the `setLookingAtPoint` instead. It
allows to use the current heading and pitch angles.

Closes: https://github.com/qgis/QGIS/issues/47935