qgscamerapose: Fix pitch initial value to avoid rotation computation error
This is a follow up from commit
b173af9a2e5097b7542668c640acb30819563be0.

This prevents a bug in QgsCameraPose::updateCamera when updating
camera rotation.
Indeed, with a mPitchAngle < 0.2, QQuaternion::fromEulerAngles(
mPitchAngle, mHeadingAngle, 0 ) will return bad rotation angle.

This is fixed is Qt6, but it has not been backported to Qt5.
See: https://bugreports.qt.io/browse/QTBUG-72103

This change is mostly useful when the `setViewFromTop` method is
called because it uses the default QgsCameraPose constructor and
therefore, the initial pitch angle is used.