qgs3dmapscene: keep track of the layers associated with the terrain
On each layer change, the `onLayersChanged` call ensures to remove the
entities of the removed layers and to create an entity for each new
layer. It uses the `mLayerEntities` container to have a list of the
existing entities. However, this list does not contain the layers
associated with the terrain.

This issues if fixed by using a new list: `mLayers` which contains all
the layers (the entities and the terrain).